+++ packages/base/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [5.0.27]
+### Added
+  - add `name` property to Worker for better debugging experience
+  
+## [5.0.27]
 ### Changed
 - `argv.getServerConfiguration` will transform `blobStore.path` & `blobStore.tempPath` to absolute path
  If path is relative it will be transformed to absolute starting from `process.configPath`.